+++ lib/src/analyzers/unnecessary_nullable_analyzer/unnecessary_nullable_analysis_config.dart
@@ -3,10 +3,6 @@ import 'package:glob/glob.dart';
 /// Represents converted unused code config which contains parsed entities.
 class UnnecessaryNullableAnalysisConfig {
   final Iterable<Glob> globalExcludes;
-  final Iterable<Glob> analyzerExcludedPatterns;
 
-  const UnnecessaryNullableAnalysisConfig(
-    this.globalExcludes,
-    this.analyzerExcludedPatterns,
-  );
+  const UnnecessaryNullableAnalysisConfig(this.globalExcludes);
 }