+++ lib/src/analyzers/unused_code_analyzer/unused_code_analysis_config.dart
@@ -3,10 +3,6 @@ import 'package:glob/glob.dart';
 /// Represents converted unused code config which contains parsed entities.
 class UnusedCodeAnalysisConfig {
   final Iterable<Glob> globalExcludes;
-  final Iterable<Glob> analyzerExcludedPatterns;
 
-  const UnusedCodeAnalysisConfig(
-    this.globalExcludes,
-    this.analyzerExcludedPatterns,
-  );
+  const UnusedCodeAnalysisConfig(this.globalExcludes);
 }