+++ lib/src/utils/analyzer_utils.dart
@@ -3,6 +3,9 @@ import 'dart:io';
 
 import 'package:analyzer/dart/analysis/analysis_context.dart';
 import 'package:analyzer/dart/analysis/analysis_context_collection.dart';
+import 'package:analyzer/dart/analysis/context_locator.dart';
+import 'package:analyzer/file_system/file_system.dart' hide File;
+import 'package:analyzer/file_system/overlay_file_system.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:analyzer/src/dart/analysis/analysis_context_collection.dart';
 import 'package:analyzer/src/dart/analysis/byte_store.dart';
@@ -18,7 +21,7 @@ AnalysisContextCollection createAnalysisContextCollection(
   String rootFolder,
   String? sdkPath,
 ) {
-  final resourceProvider = PhysicalResourceProvider.INSTANCE;
+  final resourceProvider = _prepareAnalysisOptions();
 
   return AnalysisContextCollectionImpl(
     sdkPath: sdkPath,
@@ -48,7 +51,7 @@ Set<String> getFilePaths(
 
 /// If the state location can be accessed, return the file byte store,
 /// otherwise return the memory byte store.
-ByteStore createByteStore(PhysicalResourceProvider resourceProvider) {
+ByteStore createByteStore(ResourceProvider resourceProvider) {
   const miB = 1024 * 1024 /*1 MiB*/;
   const giB = 1024 * 1024 * 1024 /*1 GiB*/;
 
@@ -86,3 +89,25 @@ Set<String> _extractDartFilesFromFolders(
                 ))
             .map((entity) => normalize(entity.path)))
         .toSet();
+
+ResourceProvider _prepareAnalysisOptions() {
+  final resourceProvider =
+      OverlayResourceProvider(PhysicalResourceProvider.INSTANCE);
+
+  final contextLocator = ContextLocator(resourceProvider: resourceProvider);
+  final roots =
+      contextLocator.locateRoots(includedPaths: [], excludedPaths: []);
+
+  for (final root in roots) {
+    final path = root.optionsFile?.path;
+    if (path != null) {
+      resourceProvider.setOverlay(
+        path,
+        content: '',
+        modificationStamp: DateTime.now().millisecondsSinceEpoch,
+      );
+    }
+  }
+
+  return resourceProvider;
+}